Added boundaries for ship's movement. Consistent movement speed in all directions
Added logic to support the movement of the ship only within the borders of the screen.

Fixed an issue in previous implementation of the movement where while moving diagonally, the ship would move faster than it would while just moving horizontally or vertically. This was cause by movement vector not being normalized, so the magnitude of said vector would be higher than one while moving diagonally