Refactored Enemy to implement component pattern. Created Enemy factory
In this commit I solved a design issue where we wanted to have multiple types of enemies in the game, but we didn't want to use inheritance.
The way i designed the solution is by implementing component pattern so we can slot in different components into different enemy types, creating a modular and scalable solution

The other pattern that was implemented was Factory Method pattern. The factory is responsible for creating an enemy based on the desired type provided as a parameter.