Introduced Particle System. Fixed collision bug
Implemented particle system. This system can now render an asset for a specified period of time in the specified location in the world, before destroying it. Used to play explosion particle effect when ship or enemy is hit

Fixed a bug in collision system where if an object in scene didn't belong to a single collision channel, it would still try to collide