Implemented State Machine in order to handle transitions between menu and playing
When we start the game, we are in the main menu state. From there we can enter play state by pressing spacebar. Once the game is over, either the ship is destroyed or the player has beaten all of the levels, then we get back to the main menu state